+++ app/Helper.php
@@ -734,4 +734,35 @@ class Helper {
 	public static function get_github_documentation_link(): string {
 		return 'https://github.com/threadi/wp-personio-integration-light/tree/master/doc';
 	}
+
+	/**
+	 * Return whether we should load styles depending on actual called backend page.
+	 *
+	 * @param string $hook The used hook.
+	 *
+	 * @return bool
+	 */
+	public static function do_not_load_styles( string $hook ): bool {
+		// bail if function is used in frontend.
+		if( ! is_admin() ) {
+			return false;
+		}
+
+		// do not load our files outside our own backend pages.
+		if( in_array( $hook, array( 'post.php', 'edit-tags.php'), true ) && function_exists( 'get_current_screen' ) ) {
+			$screen = get_current_screen();
+			if( ! in_array( $screen->post_type, apply_filters( 'personio_integration_light_do_not_load_on_cpt', array( PersonioPosition::get_instance()->get_name() ) ), true ) ) {
+				return true;
+			}
+		}
+		else {
+			// bail if no personio page is used.
+			if ( ! str_contains( $hook, 'personio' ) && ! str_contains( $hook, 'options-permalink.php' ) ) {
+				return true;
+			}
+		}
+
+		// return false to not prevent the loading of styles in backend.
+		return false;
+	}
 }